requirements.txt
asgiref==3.9.1
Django
django-cors-headers
django-tenants
django-environ==0.12.0
djangorestframework
djangorestframework-simplejwt
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
whitenoise==6.9.0
pytest==7.2.1
pytest-django
